# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: CCNA Learning SaaS API foundation
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /practice/history:
    get:
      operationId: getPracticeHistory
      tags: [practice]
      summary: List authenticated user's practice attempts
      description: Returns only the authenticated user's attempts, newest first, using cursor-based pagination. Display details are opt-in and use the topic recorded at submission, even when the question moves or either record is disabled.
      security:
        - clerkSession: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque pagination cursor returned by a previous response.
        - name: includeDetails
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Include nullable questionCode and topicName display fields. Omitted or false preserves the original response.
      responses:
        "200":
          description: Paginated practice attempt history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeHistoryResponse"
        "400":
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Practice history could not be fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /practice/topics/{topicId}/question:
    get:
      operationId: getPracticeQuestion
      tags: [practice]
      summary: Fetch a practice question for an active topic
      description: Returns one random ACTIVE question from an ACTIVE topic. Answer keys, explanations, and author-only fields are excluded.
      security:
        - clerkSession: []
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Safe practice question, before answer submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeQuestion"
        "401":
          description: No authenticated Clerk session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Topic is not active or has no active questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Practice question could not be fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /practice/questions/{questionId}/answer:
    post:
      operationId: submitPracticeAnswer
      tags: [practice]
      summary: Submit and grade an answer to an active practice question
      description: Validates that the selected option belongs to the ACTIVE question in an ACTIVE topic, then returns correctness and the answer explanation. Successful submissions are persisted as practice attempts for the authenticated user.
      security:
        - clerkSession: []
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeAnswerInput"
      responses:
        "200":
          description: Server-graded answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeAnswerResult"
        "400":
          description: Invalid request or selected option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: No authenticated Clerk session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account is disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Question is not active or its topic is not active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Practice answer could not be submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
      description: Clerk's same-origin session cookie; login and logout are managed by Clerk UI.
  schemas:
...
    PracticeTopic:
      type: object
      additionalProperties: false
      required: [id, slug, name]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
    PracticeQuestionOption:
      type: object
      additionalProperties: false
      required: [optionKey, text]
      properties:
        optionKey:
          type: string
        text:
          type: string
    PracticeQuestion:
      type: object
      additionalProperties: false
      required: [id, questionCode, topic, text, type, difficulty, options]
      properties:
        id:
          type: string
          format: uuid
        questionCode:
          type: string
        topic:
          $ref: "#/components/schemas/PracticeTopic"
        text:
          type: string
        type:
          type: string
          enum: [MULTIPLE_CHOICE_SINGLE]
        difficulty:
          type: string
          enum: [EASY, MEDIUM, HARD]
        options:
          type: array
          items:
            $ref: "#/components/schemas/PracticeQuestionOption"
    PracticeHistoryItem:
      type: object
      additionalProperties: false
      required: [id, questionId, topicId, selectedOptionKey, isCorrect, submittedAt]
      properties:
        id:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
        topicId:
          type: string
          format: uuid
        selectedOptionKey:
          type: string
        isCorrect:
          type: boolean
        submittedAt:
          type: string
          format: date-time
        questionCode:
          type: string
          nullable: true
          description: Present only with includeDetails=true. Current question code; null if unavailable.
        topicName:
          type: string
          nullable: true
          description: Present only with includeDetails=true. Current name of the topic recorded in the attempt, not the question's current topic; null if unavailable.
    PracticeHistoryResponse:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PracticeHistoryItem"
        nextCursor:
          type: string
          nullable: true
    PracticeAnswerInput:
      type: object
      additionalProperties: false
      required: [optionKey]
      properties:
        optionKey:
          type: string
          minLength: 1
          maxLength: 8
    PracticeCorrectOption:
      type: object
      additionalProperties: false
      required: [optionKey, text]
      properties:
        optionKey:
          type: string
        text:
          type: string
    PracticeAnswerResult:
      type: object
      additionalProperties: false
      required: [isCorrect, correctOption, explanation]
      properties:
        isCorrect:
          type: boolean
        correctOption:
          $ref: "#/components/schemas/PracticeCorrectOption"
        explanation:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string